==> 02-prompt.ipynb <==
import pandas as pd

input_filepath = "/Users/scottwillis/Documents/Dev/hiring-challenge-ml/docs_df/data.csv"

df = pd.read_csv(input_filepath)
df

from req_analysis.util.openai import OpenAIClient
from dotenv import load_dotenv
import os
from tqdm import tqdm

load_dotenv()

oai_client = OpenAIClient(os.getenv("AZ_OPENAI_API_KEY"), os.getenv("OPENAI_API_BASE"))

system_prompt = """/
Classify the following information:

Filename: [filename]
Section: [section]
Subsection: [subsection]
Body: [body]

Please classify the following using the provided labels:

Type (choose one):
- requirement: A set regulatory requirement that has to be applied to the regulated system.
- note: A note or additional descriptions & examples which do not contain a requirement.
- definition: A definition of a term or concept.
- section_title: A title of a section, subsection, or chapter.

Subject Matter (choose one or more):
- process_qms: Processes or the Quality Management System as a whole
- materials: Materials
- sterilization: Sterilization
- maintenance: Maintenance
- software: Software
- usability: Usability
- risk_management: Risk Management
- clinical_evaluation: Clinical Evaluation
- biocompatibility: Biocompatibility
- packaging: Packaging
- labeling: Labeling
- registration: Registration
- testing: Testing
/"""

deployment_id = "GPT-4-TURBO"
output_filepath = "/Users/scottwillis/Documents/Dev/hiring-challenge-ml/docs_df/data_oai.csv"

df['Response'] = None

# Create a progress bar
progress_bar = tqdm(total=len(df), desc="Processing")

for index, row in df.iterrows():
    filename = row["Filename"]
    section = row["Section"]
    subsection = row["Subsection"]
    body = row["Body"]

    initial_prompt = f"""/
Filename: {filename}
Section: {section}
Subsection: {subsection}
Body: {body}
/"""

    messages = oai_client.make_messages(system_prompt, initial_prompt)
    response = oai_client.call_llm(messages, deployment_id)

    df.at[index, 'Response'] = response

    # Update the progress bar
    progress_bar.update(1)

    # Save the DataFrame to a CSV file
    df.to_csv(output_filepath, index=False)

# Close the progress bar
progress_bar.close()
