==> 03-response.ipynb <==
import pandas as pd

input_filepath = "/Users/scottwillis/Documents/Dev/hiring-challenge-ml/docs_df/data_oai.csv"
df = pd.read_csv(input_filepath)
df = df[df['Response'].notna()]

def extract_content(x):
    try:
        content = x.split("content='")[1].split("'")[0]
        return content
    except IndexError:
        return None

df['Response'] = df['Response'].apply(extract_content)
df = df.dropna(subset=['Response'])

print(df)

df.iloc[0]['Response']

import re

TYPE = { # single choice
  "requirement": "A set regulatory requirement that has to be applied to the regulated system.",
  "note": "A note or additional descriptions & examples which do not contain a requirement.",
  "definition": "A definition of a term or concept.",
  "section_title": "A title of a section, subsection or chapter.",
}

SUBJECT_MATTERS = { # multiple choice
  "process_qms": "Processes or the Quality Management System as a whole",
  "materials": "Materials",
  "sterilization": "Sterilization",
  "maintenance": "Maintenance",
  "software": "Software",
  "usability": "Usability",
  "risk_management": "Risk Management",
  "clinical_evaluation": "Clinical Evaluation",
  "biocompatibility": "Biocompatibility",
  "packaging": "Packaging",
  "labeling": "Labeling",
  "registration": "Registration",
  "testing": "Testing",
}

def extract_type(response):
    type_match = re.search(r'Type: ([a-zA-Z_]+)', response)
    if type_match:
        type_value = type_match.group(1)
        # Validate and filter type option
        if type_value in TYPE:
            return type_value
    return None

def extract_subject_matter(response):
    subject_matter_match = re.search(r'Subject Matter: (.+)', response)
    if subject_matter_match:
        subject_matter_value = subject_matter_match.group(1).split(', ')
        # Validate and filter subject matter options
        subject_matter_value = [option for option in subject_matter_value if option in SUBJECT_MATTERS]
        if subject_matter_value:
            return subject_matter_value
    return []

# Apply the custom functions to extract Type and Subject Matter for each row
df['Type'] = df['Response'].apply(lambda x: extract_type(x))
df['Subject Matter'] = df['Response'].apply(lambda x: extract_subject_matter(x))

# Print the updated DataFrame
print(df.iloc[0])

output_filepath = "/Users/scottwillis/Documents/Dev/hiring-challenge-ml/docs_df/data_oai_parsed.csv"
df.to_csv(output_filepath)
